+++ spec/testLoadUseBundle.js
@@ -10,6 +10,9 @@ const loggerFactory = tsl.LFService.createLoggerFactory();
 
 // Get a logger called "Hello"
 const loggerOld = loggerFactory.getLogger("Hello");
+loggerOld.error("Log4j on error", new Error("log4j error"));
+
+tsl.CategoryServiceFactory.setDefaultConfiguration(new tsl.CategoryDefaultConfiguration(tsl.LogLevel.Debug));
 
 const catRoot1 = new tsl.Category("root1", null);
 const catChild = new tsl.Category("r1child1", catRoot1);
@@ -21,8 +24,17 @@ const r2Child2 = new tsl.Category("r2child2", catRoot2);
 
 const logger = tsl.CategoryServiceFactory.getLogger(catRoot1);
 logger.error("Error in normal console", new Error("fail"), catRoot1);
+logger.debug("This is on debug");
 
 const logger2 = tsl.CategoryServiceFactory.getLogger(catRoot2);
 logger2.errorc(function() { return "failed"; }, function() { return new Error("oops"); });
+logger2.debugc(function() { return "category on debug"; });
 
-loggerOld.error("Log4j on error", new Error("log4j error"));
+/*
+CategoryServiceFactory.setDefaultConfiguration(new CategoryDefaultConfiguration(LogLevel.Debug));
+
+export const appLog: Category = new Category('myApp');
+export const log: CategoryLogger = CategoryServiceFactory.getLogger(appLog);
+
+log.debug('I see it!!!');
+*/